Finished first improved image get handling
The API has been simplified to accept an array of format
short names that describe a preference, which image to get.
The image is constructed and returned, later a function can
be called to write out that image into a buffer. It could
also be manuipulated in the user's application using the
image crate's functions.